# docs/api/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
...
paths:
...
  /api/v1/searchs/{type}:
    get:
      tags:
      - Search
      summary: Search entities
      description: "Performs a search across Diamond Analytics entities based on the\
        \ provided type. Supported types are: 'stadium', 'team', and 'player'. When\
        \ searching for players, the 'playerType' parameter must be specified as either\
        \ 'position' or 'pitcher'. Results are returned in a paginated format."
      operationId: searchStadium
      parameters:
      - name: type
        in: path
        required: true
        schema:
          type: string
      - name: query
        in: query
        required: true
        schema:
          type: string
      - name: playerType
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "400":
          description: "Missing or invalid parameters (e.g., playerType not provided\
            \ or incorrect)"
          content:
            application/json: {}
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: "Invalid search type (not stadium, team, or player)"
          content:
            application/json: {}
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Successfully retrieved search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedModel"
...
  /api/v1/cache:
    get:
      tags:
      - Cache Management
      summary: Get all cache names
      description: Retrieve a list of all active cache names managed by the application.
      operationId: getCaches
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "401":
          description: Unauthorized access
          content:
            application/json: {}
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Successfully retrieved cache names
          content:
            application/json:
              schema:
                type: string
    delete:
      tags:
      - Cache Management
      summary: Clear all caches
      description: Perform a global purge of all application caches.
      operationId: clearAllCaches
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "401":
          description: Unauthorized access
          content:
            application/json: {}
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Successfully cleared all caches
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /api/v1/cache/{name}:
    delete:
      tags:
      - Cache Management
      summary: Clear a specific cache
      description: Clear the contents of a single cache.
      operationId: clearSingleCache
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Cache not found
          content:
            application/json: {}
        "401":
          description: Unauthorized access
          content:
            application/json: {}
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Successfully cleared the cache
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
components:
  schemas:
...
    AuthResponse:
      type: object
      properties:
        status:
          type: string
          enum:
          - SUCCESS
          - FAILURE
        message:
          type: string
        error:
          type: string
...
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
    PagedModel:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        page:
          $ref: "#/components/schemas/PageMetadata"
...